# openAPI_Spec/admin端接口.openapi.yaml
openapi: 3.1.0
info:
  title: admin端接口
  description: ''
  version: 1.0.0
...
paths:
...
  /admin/dish:
    put:
      summary: 修改菜品
      deprecated: false
      description: ''
      tags:
        - 菜品相关接口
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                categoryId:
                  type: integer
                  format: int64
                description:
                  type: string
                flavors:
                  type: array
                  items:
                    type: object
                    properties:
                      dishId:
                        type: integer
                        format: int64
                      id:
                        type: integer
                        format: int64
                      name:
                        type: string
                      value:
                        type: string
                    title: DishFlavor
                    $$ref: '#/definitions/DishFlavor'
                    required:
                      - name
                      - value
                id:
                  type: integer
                  format: int64
                image:
                  type: string
                name:
                  type: string
                price:
                  type: number
                status:
                  type: integer
                  format: int32
              title: DishDTO
              $$ref: '#/definitions/DishDTO'
              required:
                - categoryId
                - id
                - image
                - name
                - price
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: string
                  msg:
                    type: string
                title: R«string»
                $$ref: '#/definitions/R«string»'
                required:
                  - code
          headers: {}
      security: []
    delete:
      summary: 批量删除菜品
      deprecated: false
      description: ''
      tags:
        - 菜品相关接口
      parameters:
        - name: ids
          in: query
          description: 菜品id，之间用逗号分隔
          required: true
          example: 1,2,3
          schema:
            type: string
      requestBody:
        content:
          text/plain:
            schema:
              type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: string
                  msg:
                    type: string
                title: R«string»
                $$ref: '#/definitions/R«string»'
                required:
                  - code
          headers: {}
      security: []
    post:
      summary: 新增菜品
      deprecated: false
      description: ''
      tags:
        - 菜品相关接口
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                categoryId:
                  type: integer
                  format: int64
                  description: 分类id
                description:
                  type: string
                  description: 菜品描述
                flavors:
                  type: array
                  items:
                    type: object
                    properties:
                      dishId:
                        type: integer
                        format: int64
                        description: 菜品id
                      id:
                        type: integer
                        format: int64
                        description: 口味id
                      name:
                        type: string
                        description: 口味名称
                      value:
                        type: string
                        description: 口味值
                    title: DishFlavor
                    $$ref: '#/definitions/DishFlavor'
                    required:
                      - name
                      - value
                  description: 口味
                id:
                  type: integer
                  format: int64
                  description: 菜品id
                image:
                  type: string
                  description: 菜品图片路径
                name:
                  type: string
                  description: 菜品名称
                price:
                  type: number
                  description: 菜品价格
                status:
                  type: integer
                  format: int32
                  description: 菜品状态：1为起售，0为停售
              title: DishDTO
              $$ref: '#/definitions/DishDTO'
              required:
                - categoryId
                - image
                - name
                - price
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: string
                  msg:
                    type: string
                title: R«string»
                $$ref: '#/definitions/R«string»'
                required:
                  - code
          headers: {}
      security: []
  /admin/dish/{id}:
    get:
      summary: 根据id查询菜品
      deprecated: false
      description: ''
      tags:
        - 菜品相关接口
      parameters:
        - name: id
          in: path
          description: 菜品id
          required: true
          example: '101'
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: object
                    properties:
                      categoryId:
                        type: integer
                        format: int64
                      categoryName:
                        type: string
                      description:
                        type: string
                      flavors:
                        type: array
                        items:
                          type: object
                          properties:
                            dishId:
                              type: integer
                              format: int64
                            id:
                              type: integer
                              format: int64
                            name:
                              type: string
                            value:
                              type: string
                          title: DishFlavor
                          $$ref: '#/definitions/DishFlavor'
                          required:
                            - dishId
                            - id
                            - name
                            - value
                      id:
                        type: integer
                        format: int64
                      image:
                        type: string
                      name:
                        type: string
                      price:
                        type: number
                      status:
                        type: integer
                        format: int32
                      updateTime:
                        type: string
                        format: date-time
                    title: DishVO
                    $$ref: '#/definitions/DishVO'
                    required:
                      - categoryId
                      - categoryName
                      - flavors
                      - description
                      - id
                      - image
                      - name
                      - price
                      - status
                      - updateTime
                  msg:
                    type: string
                title: R«DishVO»
                $$ref: '#/definitions/R«DishVO»'
                required:
                  - code
                  - data
          headers: {}
      security: []
  /admin/dish/list:
    get:
      summary: 根据分类id查询菜品
      deprecated: false
      description: ''
      tags:
        - 菜品相关接口
      parameters:
        - name: categoryId
          in: query
          description: 分类id
          required: true
          example: '101'
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        categoryId:
                          type: integer
                          format: int64
                        createTime:
                          type: string
                          format: date-time
                        createUser:
                          type: integer
                          format: int64
                        description:
                          type: string
                        id:
                          type: integer
                          format: int64
                        image:
                          type: string
                        name:
                          type: string
                        price:
                          type: number
                        status:
                          type: integer
                          format: int32
                        updateTime:
                          type: string
                          format: date-time
                        updateUser:
                          type: integer
                          format: int64
                      title: Dish
                      $$ref: '#/definitions/Dish'
                  msg:
                    type: string
                title: R«List«Dish»»
                $$ref: '#/definitions/R«List«Dish»»'
                required:
                  - code
          headers: {}
      security: []
  /admin/dish/page:
    get:
      summary: 菜品分页查询
      deprecated: false
      description: ''
      tags:
        - 菜品相关接口
      parameters:
        - name: categoryId
          in: query
          description: 分类id
          required: false
          example: '101'
          schema:
            type: string
        - name: name
          in: query
          description: 菜品名称
          required: false
          example: 宫保鸡丁
          schema:
            type: string
        - name: page
          in: query
          description: 页码
          required: true
          example: '1'
          schema:
            type: string
        - name: pageSize
          in: query
          description: 每页记录数
          required: true
          example: '10'
          schema:
            type: string
        - name: status
          in: query
          description: 分类状态
          required: false
          example: '1'
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: number
                  msg:
                    type: 'null'
                  data:
                    type: object
                    properties:
                      total:
                        type: number
                        description: 总记录数
                      records:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: number
                            name:
                              type: string
                            categoryId:
                              type: number
                            price:
                              type: number
                            image:
                              type: string
                            description:
                              type: string
                            status:
                              type: number
                            updateTime:
                              type: string
                            categoryName:
                              type: string
                              description: 分类名称
                          required:
                            - id
                            - name
                            - categoryId
                            - price
                            - image
                            - description
                            - status
                            - updateTime
                            - categoryName
                        description: 当前页数据
                    required:
                      - total
                      - records
                required:
                  - code
          headers: {}
      security: []
  /admin/dish/status/{status}:
    post:
      summary: 菜品起售、停售
      deprecated: false
      description: ''
      tags:
        - 菜品相关接口
      parameters:
        - name: status
          in: path
          description: 菜品状态：1为起售，0为停售
          required: true
          example: '1'
          schema:
            type: string
        - name: id
          in: query
          description: 菜品id
          required: true
          example: '101'
          schema:
            type: string
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: string
                  msg:
                    type: string
                title: R«string»
                $$ref: '#/definitions/R«string»'
                required:
                  - code
          headers: {}
      security: []
...
components:
...
  securitySchemes: {}
servers: []
security: []